+++ modules/AWSPowerShell/Cmdlets/AmazonRDS/Basic/New-RDSDBClusterParameterGroup-Cmdlet.cs
@@ -83,7 +83,7 @@ namespace Amazon.PowerShell.Cmdlets.RDS
         /// <para>The DB cluster parameter group family name. A DB cluster parameter group can be associated
         /// with one and only one DB cluster parameter group family, and can be applied only to
         /// a DB cluster running a database engine and engine version compatible with that DB
-        /// cluster parameter group family.</para>
+        /// cluster parameter group family.</para><para><b>Aurora MySQL</b></para><para>Example: <code>aurora5.6</code>, <code>aurora-mysql5.7</code></para><para><b>Aurora PostgreSQL</b></para><para>Example: <code>aurora-postgresql9.6</code></para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]